+++ test/hypermedia_feature_test.rb
@@ -131,7 +131,7 @@ class HypermediaTest < MiniTest::Spec
       end
 
       it "sets up an empty link list if no links found in the document" do
-        @bookmarks_with_links.from_xml(%{<bookmarks/>}).links.must_equal({})
+        @bookmarks_with_links.from_xml(%{<bookmarks/>}).links.must_equal nil
       end
     end