+++ CHANGES.markdown
@@ -4,6 +4,7 @@
 
 * `Hyperlink#to_hash` now returns stringified keys.
 * Removed `Representer#before_serialize` hook.
+* Represented#links now returns `nil` when no parsing has happened.
 
 ## Internals